Crate hdfs

fs-hdfs is a library for accessing to HDFS cluster. Basically, it provides libhdfs FFI APIs. It also provides more idiomatic and abstract Rust APIs, hiding manual memory management and some thread-safety problem of libhdfs. Rust APIs are highly recommended for most users.

Important Note

The original libhdfs implementation allows only one HdfsFs instance for the same namenode because libhdfs only keeps a single hdfsFs entry for each namenode. As a result, a global cache HDFS_CACHE is introduced to control only one single hdfsFs entry created for each namenode. Contrast, HdfsFs instance itself is thread-safe.

Usage

in Cargo.toml:

[dependencies]
fs-hdfs = "0.1.1"

or

[dependencies.fs-hdfs]
git = "https://github.com/yahoNanJing/fs-hdfs"

fs-hdfs uses libhdfs. Firstly, we need to add library path to find the libhdfs. An example for MacOS,

export DYLD_LIBRARY_PATH=$HADOOP_HOME/lib/native:$JAVA_HOME/jre/lib/server

Here, $HADOOP_HOME and $JAVA_HOME need to be specified and exported.

Since our dependent libhdfs is JNI native implementation, it requires the proper CLASSPATH. An example,

export CLASSPATH=$CLASSPATH:`hadoop classpath`

Testing

The test also requires the CLASSPATH. In case that the java class of org.junit.Assert can’t be found. Refine the $CLASSPATH as follows:

export CLASSPATH=$CLASSPATH:`hadoop classpath`:$HADOOP_HOME/share/hadoop/tools/lib/*

Then you can run

cargo test

Example

let fs: HdfsFs = HdfsFs::new("hdfs://localhost:8020/").ok().unwrap();
match fs.mkdir("/data") {
  Ok(_) => { println!("/data has been created") },
  Err(_)  => { panic!("/data creation has failed") }
};

Modules

err

hdfs

Rust APIs wrapping libhdfs API, providing better semantic and abstraction it’s a modified version of hdfs-rs

minidfs

Mainly for unit test MiniDfs Cluster