Module hdfs::binding

libhdfs binding APIs

Structs

MiniDfsConf
Represents a configuration to use for creating a Native MiniDFSCluster

hdfsFileInfo

hdfsReadStatistics
Information about a file/directory.

Enums

NativeMiniDfsCluster
Opaque Pointer for NativeMiniDfsCluster

hadoopRzBuffer
Opaque Pointer of hadoopRzBuffer

hadoopRzOptions
Opaque Pointer of hadoopRzOptions

hdfsBuilder
Opaque Pointer of hdfsBuilder

hdfsFS
Opaque Pointer of hdfsFS

hdfsFile
Opaque Pointer of hdfsFile

tObjectKind

Functions

hadoopReadZero
Perform a byte buffer read. If possible, this will be a zero-copy (mmap) read.

hadoopRzBufferFree
Release a buffer obtained through readZero.

hadoopRzBufferGet
Get a pointer to the raw buffer returned from readZero.

hadoopRzBufferLength
Determine the length of the buffer returned from readZero.

hadoopRzOptionsAlloc
Allocate a zero-copy options structure.

hadoopRzOptionsFree
Free a hadoopRzOptionsFree structure.

hadoopRzOptionsSetByteBufferPool
Set the ByteBufferPool to use with read0.

hadoopRzOptionsSetSkipChecksum
Determine whether we should skip checksums in read0.

hdfsAvailable
Number of bytes that can be read from this input stream without blocking.

hdfsBuilderConfSetStr
Set a configuration string for an HdfsBuilder.

hdfsBuilderConnect
Connect to HDFS using the parameters defined by the builder.

hdfsBuilderSetForceNewInstance
Force the builder to always create a new instance of the FileSystem, rather than possibly finding one in the cache.

hdfsBuilderSetKerbTicketCachePath
Set the path to the Kerberos ticket cache to use when connecting to the HDFS cluster.

hdfsBuilderSetNameNode
Set the HDFS NameNode to connect to.

hdfsBuilderSetNameNodePort
Set the port of the HDFS NameNode to connect to.

hdfsBuilderSetUserName
Set the username to use when connecting to the HDFS cluster.

hdfsChmod
hdfsChmod

hdfsChown
Change the user and/or group of a file or directory.

hdfsCloseFile
Close an open file.

hdfsConfGetInt
Get a configuration integer.

hdfsConfGetStr
Get a configuration string.

hdfsConfStrFree
Free a configuration string found with hdfsConfGetStr.

hdfsConnect
Connect to a hdfs file system.

hdfsConnectAsUser
Connect to a hdfs file system as a specific user.

hdfsConnectAsUserNewInstance
Connect to an hdfs file system.

hdfsConnectNewInstance
Connect to an hdfs file system.

hdfsCopy
Copy file from one filesystem to another.

hdfsCreateDirectory
Make the given file and all non-existent parents into directories.

hdfsDelete
Delete file.

hdfsDisconnect
hdfsDisconnect - Disconnect from the hdfs file system. Disconnect from hdfs.

hdfsExists
Checks if a given path exsits on the filesystem

hdfsFileFreeReadStatistics
Free some HDFS read statistics.

hdfsFileGetReadStatistics
Get read statistics about a file. This is only applicable to files opened for reading.

hdfsFileIsEncrypted
hdfsFileIsEncrypted: determine if a file is encrypted based on its hdfsFileInfo.

hdfsFileIsOpenForRead
Determine if a file is open for read.

hdfsFileIsOpenForWrite
Determine if a file is open for write.

hdfsFlush
Flush the data.

hdfsFreeBuilder
Free an HDFS builder.

hdfsFreeFileInfo
Free up the hdfsFileInfo array (including fields)

hdfsFreeHosts
Free up the structure returned by hdfsGetHosts

hdfsGetCapacity
Return the raw capacity of the filesystem.

hdfsGetDefaultBlockSize
Get the default blocksize.

hdfsGetDefaultBlockSizeAtPath
Get the default blocksize at the filesystem indicated by a given path.

hdfsGetHosts
Get hostnames where a particular block (determined by pos & blocksize) of a file is stored. The last element in the array is NULL. Due to replication, a single block could be present on multiple hosts.

hdfsGetPathInfo
Get information about a path as a (dynamically allocated) single hdfsFileInfo struct. hdfsFreeFileInfo should be called when the pointer is no longer needed.

hdfsGetUsed
Return the total raw size of all files in the filesystem.

hdfsGetWorkingDirectory
Get the current working directory for the given filesystem.

hdfsHFlush
Flush out the data in client's user buffer. After the return of this call, new readers will see the data.

hdfsHSync
Similar to posix fsync, Flush out the data in client's user buffer. all the way to the disk device (but the disk may have it in its cache).

hdfsListDirectory
Get list of files/directories for a given directory-path. hdfsFreeFileInfo should be called to deallocate memory.

hdfsMove
Move file from one filesystem to another.

hdfsNewBuilder
Create an HDFS builder.

hdfsOpenFile
Open a hdfs file in given mode.

hdfsPread
Positional read of data from an open file.

hdfsRead
Read data from an open file.

hdfsReadStatisticsGetRemoteBytesRead
HDFS read statistics for a file,

hdfsRename
Rename file.

hdfsSeek
Seek to given offset in file.

hdfsSetReplication
Set the replication of the specified file to the supplied value

hdfsSetWorkingDirectory
Set the working directory. All relative paths will be resolved relative to it.

hdfsTell
Get the current offset in the file, in bytes.

hdfsUtime
hdfsUtime

hdfsWrite
Write data into an open file.

nmdConfigureHdfsBuilder
Configure the HDFS builder appropriately to connect to this cluster.

nmdCreate
Create a NativeMiniDfsCluster

nmdFree
Destroy a Native MiniDFSCluster

nmdGetNameNodeHttpAddress
Get the http address that's in use by the given (non-HA) nativeMiniDfs

nmdGetNameNodePort
Get the port that's in use by the given (non-HA) nativeMiniDfs

nmdShutdown
Shut down a NativeMiniDFS cluster

nmdWaitClusterUp
Wait until a MiniDFSCluster comes out of safe mode.

Type Definitions

tOffset
offset within the file

tPort
port

tSize
size of data for read/write io ops

tTime
time type in seconds