Struct hdfs::binding::MiniDfsConf

pub struct MiniDfsConf {
    // some fields omitted
}

Represents a configuration to use for creating a Native MiniDFSCluster

Methods

impl MiniDfsConf

fn new() -> MiniDfsConf

fn set_do_format(&mut self, on: bool) -> &mut MiniDfsConf

Set TRUE if the cluster should be formatted prior to startup

fn do_format(&self) -> bool

The cluster will be formatted prior to startup if TRUE

fn set_web_hdfs(&mut self, enable: bool) -> &mut MiniDfsConf

Set TRUE in order to enable webhdfs in MiniDfsCluster

fn web_hdfs_enabled(&self) -> bool

webhdfs in MiniDfsCluster will be available if TRUE

fn set_http_port(&mut self, port: i32) -> &mut MiniDfsConf

Set http port of the namenode in MiniDfsCluster

fn http_port(&self) -> i32

The http port of the namenode in MiniDfsCluster

fn set_short_circuit(&mut self, enable: bool) -> &mut MiniDfsConf

Set TRUE if we should configure short circuit.

fn short_circuit_enabled(&self) -> bool

short circuit will be available if TRUE