Struct hdfs::hdfs::HdfsFs

pub struct HdfsFs { /* fields omitted */ }

Hdfs Filesystem

It is basically thread safe because the native API for hdfsFs is thread-safe.

Implementations

impl HdfsFs

pub fn new(path: &str) -> Result<HdfsFs, HdfsErr>

Create instance of HdfsFs with a global cache. For each namenode uri, only one instance will be created

pub fn default() -> Result<HdfsFs, HdfsErr>

The default NameNode configuration will be used (from the XML configuration files)

pub fn remove(path: &str) -> Result<Option<HdfsFs>, HdfsErr>

pub fn url(&self) -> &str

Get HDFS namenode url

pub fn raw(&self) -> *mut hdfs_internal

Get a raw pointer of JNI API’s HdfsFs

pub fn open(&self, path: &str) -> Result<HdfsFile, HdfsErr>

open a file to read

pub fn open_with_buf_size(
&self,
path: &str,
buf_size: i32
) -> Result<HdfsFile, HdfsErr>

open a file to read with a buffer size

pub fn get_file_status(&self, path: &str) -> Result<FileStatus, HdfsErr>

Get the file status, including file size, last modified time, etc

pub fn list_status(&self, path: &str) -> Result<Vec<FileStatus>, HdfsErr>

Get the file status for each entry under the specified directory

pub fn default_blocksize(&self) -> Result<usize, HdfsErr>

Get the default blocksize.

pub fn block_size(&self, path: &str) -> Result<usize, HdfsErr>

Get the default blocksize at the filesystem indicated by a given path.

pub fn capacity(&self) -> Result<usize, HdfsErr>

Return the raw capacity of the filesystem.

pub fn used(&self) -> Result<usize, HdfsErr>

Return the total raw size of all files in the filesystem.

pub fn exist(&self, path: &str) -> bool

Checks if a given path exsits on the filesystem

pub fn get_hosts(
&self,
path: &str,
start: usize,
length: usize
) -> Result<BlockHosts, HdfsErr>

Get hostnames where a particular block (determined by pos & blocksize) of a file is stored. The last element in the array is NULL. Due to replication, a single block could be present on multiple hosts.

pub fn create(&self, path: &str) -> Result<HdfsFile, HdfsErr>

pub fn create_with_overwrite(
&self,
path: &str,
overwrite: bool
) -> Result<HdfsFile, HdfsErr>

pub fn create_with_params(
&self,
path: &str,
overwrite: bool,
buf_size: i32,
replica_num: i16,
block_size: i32
) -> Result<HdfsFile, HdfsErr>

pub fn chmod(&self, path: &str, mode: i16) -> bool

set permission

pub fn chown(&self, path: &str, owner: &str, group: &str) -> bool

pub fn append(&self, path: &str) -> Result<HdfsFile, HdfsErr>

Open a file for append

pub fn mkdir(&self, path: &str) -> Result<bool, HdfsErr>

create a directory

pub fn rename(&self, old_path: &str, new_path: &str) -> Result<bool, HdfsErr>

Rename file.

pub fn set_replication(&self, path: &str, num: i16) -> Result<bool, HdfsErr>

Set the replication of the specified file to the supplied value

pub fn delete(&self, path: &str, recursive: bool) -> Result<bool, HdfsErr>

Delete file.