Struct hdfs::native::hdfsFileInfo

pub struct hdfsFileInfo {
    pub mKind: tObjectKind,
    pub mName: *const c_char,
    pub mLastMod: tTime,
    pub mSize: tOffset,
    pub mReplication: c_short,
    pub mBlockSize: tOffset,
    pub mOwner: *const c_char,
    pub mGroup: *const c_char,
    pub mPermissions: c_short,
    pub mLastAccess: tTime,
}

Fields

mKind

file or directory

mName

the name of the file

mLastMod

the last modification time for the file in seconds

mSize

the size of the file in bytes

mReplication

the count of replicas

mBlockSize

the block size for the file

mOwner

the owner of the file

mGroup

the group associated with the file

mPermissions

the permissions associated with the file

mLastAccess

the last access time for the file in seconds