Struct hdfs::FileStatus

pub struct FileStatus<'a> {
    // some fields omitted
}

Interface that represents the client side information for a file or directory.

Methods

impl<'a> FileStatus<'a>

fn name(&self) -> &'a str

Get the name of the file

fn is_file(&self) -> bool

Is this a file?

fn is_directory(&self) -> bool

Is this a directory?

fn owner(&self) -> &'a str

Get the owner of the file

fn group(&self) -> &'a str

Get the group associated with the file

fn permission(&self) -> i16

Get the permissions associated with the file

fn len(&self) -> usize

Get the length of this file, in bytes.

fn block_size(&self) -> usize

Get the block size of the file.

fn replica_count(&self) -> i16

Get the replication factor of a file.

fn last_modified(&self) -> time_t

Get the last modification time for the file in seconds

fn last_accced(&self) -> time_t

Get the last access time for the file in seconds