Crate hdfs

hdfs-rs is a library for accessing to HDFS cluster. Basically, it provides libhdfs FFI APIs. It also provides more idiomatic and abstract Rust APIs, hiding manual memory management and some thread-safety problem of libhdfs. Rust APIs are highly recommended for most users.

Modules

minidfs
Mini HDFS Cluster for easily building unit tests MiniDfs Cluster

native
libhdfs native binding APIs libhdfs FFI Binding APIs

Structs

BlockHosts
Includes hostnames where a particular block of a file is stored.

FileStatus
Interface that represents the client side information for a file or directory.

HdfsFS
Hdfs Filesystem

HdfsFile
open hdfs file

HdfsFsCache
HdfsFs Cache. Basically, It is a cache as well as a way to guarantees thread-safe when you get HdfsFs.

HdfsUtil
Hdfs Utility

RzBuffer
A buffer returned from zero-copy read. This buffer will be automatically freed when its lifetime is finished.

RzOptions
Options for zero-copy read

Enums

HdfsErr